backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
pyjwt
python-dotenv